Requirement: Neutron DHCP servers in HA
We currently have most openstack services in HA, including
neutron-dhcp-agent service. Meaning that there are multiple neutron-dhcp-agent
processes running on the controllers. One of those is selected when
there is a request to create a dhcp server namespace. A single dhcp
server will be created for that tenant.
This issue is created as a proposal for requiring OPNFV Target System to
enable multiple dhcp servers per tenant in HA, so that a tenant would
have multiple dhcp servers in HA in case one goes down.

JIRA: GENESIS-43

Change-Id: I1386c790834c0924ec90eb505d14ca7c8449f408
Signed-off-by: Tim Rozet <trozet@redhat.com>