Requirement: SDN Controller Layer 3 forwarding support
This is a proposal to require Installers to support the ability to
enable L3 routing functionality in the SDN Controller (where possible).
Some SDN controllers may not support this functionality, in which case
it is acceptable to use Neutron L3 agent.
OpenDaylight also supports distributed virtual routing (DVR). This
requirement also proposes that Installers should be capable of enabling
DVR on SDN controllers which support DVR.

JIRA: GENESIS-44

Change-Id: Ib0e585f05cc75d30b198863bba342e2062ca255e
Signed-off-by: Tim Rozet <trozet@redhat.com>