Requirement: Ability to install with upstream artifacts
Installers should support deployment of OPNFV using artifacts (source
code, packages, etc) from 3rd-party repos (vendor specific distros and
trunk versions). For example, if an OPNFV project required a specific
version of OVS (and provided the compatible artifact) to be installed,
an installer would be able to point directly to that artifact and
build/deploy with it.

JIRA: GENESIS-39

Change-Id: Iea8eddf36b03c124f7fb048f4c301179b2697b59
Signed-off-by: Tim Rozet <trozet@redhat.com>