Requirement: Supported Layer 1/2 Networking Configurations
Installers should support multiple kinds of network configurations for
deployments. These shall include the following:
1) Multiple VLANs (untagged and tagged links)
2) NIC Bonding/Etherchannel
3) Multiple NICs
These types may also be in combination with one another. For example, if
a deployment uses 4 logical broadcast domains (VLANs) it may be split
into 4 sub-interfaces on a single interface, which is then mapped to a
bonded link consisting of 2 links.

JIRA: GENESIS-61

Change-Id: Id30cfe1c01c3b82d94aaf454d6496180c29732e9
Signed-off-by: Tim Rozet <trozet@redhat.com>