Requirement: "All-in-one" build for Brahmaputra
Installers which create a build for Brahmaputra, should create
the build as an "all-in-one" build. The the build process of
the installer creates a single entity (e.g. ISO) - which has
all the artifacts considered and required by all the projects
for Brahmaputra packaged in. Or in other terms and as an example:
If there are 4 different versions of OVS - all these 4 versions
would be contained in the "all in one build".
Note: This requirement only applies to installers which support
a "build" phase (i.e. create a bootable image, like an iso-image
from the different artifacts required).

JIRA: GENESIS-74

Change-Id: Ia6f126ae8a062c8e3158b65835a9ae2429cba321
Signed-off-by: Frank Brockners <fbrockne@cisco.com>