Key artifacts and their location
For Brahmaputra, we're using an "all in one" build
approach. There is no need for a specific configuration
file that installers would use as part of their build
process which points to the location of components that
will be included.
The "artifacts-location" file is to serve
documentation purposes only for the Brahmaputra
release. It is to list all key artifacts and their
location. This patch creates the starting point for
the documentation.

This patch is a replica of 4585 which was abandoned
due to merge/rebase issues.

Change-Id: I09dba4686d908cf25edf2ff9a4de054959e035a1
Signed-off-by: Frank Brockners <fbrockne@cisco.com>