Evolve templates to include Colorado release
Evolved templates for system and ux requirements
to cover Colorado (in addition to Brahmaputra).

Requirements that are only for Brahmaputra
are flagged as "Brahmaputra:".

Requirements that are only for Colorado
are flagged as "Colorado:". Placeholders added only,
to be used as guiding examples (see e.g. for
Openstack version or OpenDaylight version).

Requirements that apply from Brahmaputra onwards are
tagged as "Brahmpautra and later:".

System.rst files which describe the components for
a release are move to release specific directories.

Change-Id: I614921a9bf9c8166c9a19000ee822bdc3b9e2994
Signed-off-by: Frank Brockners <fbrockne@cisco.com>